Centralize configman things into a utility library
This extracts config_from_configman() into a utility library so we can use
it elsewhere in the webapp.
